+++ mopidy/backend.py
@@ -267,12 +267,42 @@ class PlaylistsProvider(object):
     # currently make available. lookup() should be used for getting full
     # playlists with all details.
 
+    def get_playlists(self, ref=True):
+        """
+        Get available playlists.
+
+        If ``ref`` is :class:`False`, a list of :class:`mopidy.models.Playlist`
+        is returned.
+
+        If ``ref`` is :class:`True`, a list of :class:`mopidy.models.Ref`
+        referencing the playlists is returned.
+
+        Implementors should use :class:`True` as the default value for the
+        ``ref`` keyword argument.
+
+        .. versionadded:: 0.20
+
+        .. deprecated:: 0.20
+            The ``ref`` keyword argument will be removed in the future together
+            with the behavior associated with ``ref`` being :class:`False`.
+        """
+        if ref:
+            return []
+        else:
+            return self.playlists
+
     @property
     def playlists(self):
         """
         Currently available playlists.
 
         Read/write. List of :class:`mopidy.models.Playlist`.
+
+        .. deprecated:: 0.20
+            If your backend requires Mopidy >= 0.20, implement
+            :meth:`get_playlists` instead of this property. If you must stay
+            compatible with Mopidy < 0.20, you're free to implement
+            :meth:`get_playlists` in addition to this property.
         """
         return []